Fix bugs in buddhabrot calculation
There were multiple issues that were causing the origin of the image
(0,0) to gain many hits from the trajectory saving procedures. One was
due to the starting of the Mandelbrot function at z = 0 + 0i, instead of
at z0.

Another issue was an off-by-one error in the number of iterations
passed to the saveTrajectory function.

Finally, the anti-buddhabrot check was incorrectly reversed.